--- share/shwap/p2p/bitswap/hasher_test.go
@@ -1,1 +1,0 @@
-package bitswap
